Add parameter in configuration for minimize to tray
